Corrected entries for Robert Murphy
VSFS student pointed out that Robert Murphy appeared to take office as
Deputy Under Secretary of State for Political Affairs before Congress
created the position. After discussing the issue with Evan, it appears
that Murphy became Deputy Under Secretary under the 1949 Department of
State Organization Act (the Department was allotted two Deputies but
does not appear to have filled both positions at the time). Murphy used
the *title* Deputy Under Secretary of State for Political Affairs, but
his appointment was authorized by the 1949 act. Therefore, I’ve moved
his 1953-1955 appointment to the “Deputy Under Secretary” page and
added notes to the entry for his service as Deputy Under Secretary and
his service as A/S for UN Affairs to clear up any latent confusion. He
now appears on the page of the position with the correct authorizing
act, with a notation that he used a different title. Whew!