David L. Gregg birthdate
David L. Gregg’s birth year comes from Pauline King, ed., “The Diaries
of David Lawrence Gregg: An American Diplomat in Hawaii” (Honolulu:
Hawaiian Historical Society, 1982), in which she writes “Those facts
that have been determined place Gregg as one of five children of Asa
and Violette Gragg. His father was born in Rindge, Cheshire County, New
Hampshire, but he left there sometime before 1814 and moved to New York
in the vicinity of Albany. At about this time, Asa began to spell the
family name with an ‘e,’ or as Gregg. David was born on May 9th,
probably in 1814 and either in Richfield, Otsego County, or Albany,
Union County.” Since this seems to be the best estimate of his birth, I
am using the 1814 year.