International Orgs/Last of Evan's updates
Completed the revamp of international orgs and also input the last of
Evan’s changes from the summer.

Joe: Evan created a new A/S-equivalent position, the Director of the
Office of Management Policy, Rightsizing, and Innovation. I created the
file based on another position file, but I wasn’t sure about some of
the elements (such as <is-dirty>). Would be grateful if you could take
a look at this when you get a chance.