Death date
Date of death provided by Princeton University. There records have him
as John Fletcher Martin, but it is undoubtedly the same person (and
Fletcher was his mother’s maiden name). His application to join the
Foreign Service, which is on the archival side in St. Louis, lists him
as John Funk Martin, Jr., and so for now I am leaving the name as is.