Add schema check for position order
This check compares the start date for each position against the end date of the previous position, raising a warning when the new position starts before the end of the previous one. For example, “Possible position ordering problem: This position started on May 12, 1989, but the previous position didn't end until afterward, July 27, 1989.” It’s raised as a warning rather than an error because for some positions we do not track exact start or end dates, and for others positions are not necessarily occupied exclusively; we could probably develop a classification for these kinds positions that would help us suppress unnecessary warnings. But hopefully this will help us avoid obvious errors like the Blaine-Frelinghuysen fixed earlier today.