error revision
initial revision to correct error reported by Aaron Marrs in 12/5/2019 email to DIDTeam. Aaron suggests further research to validate this revised termination of appointment date (from 1986 and 1991 printed editions of POCOM) since it is two weeks after Watson's successor was appointed and entered on duty. Will check cables to see if clarification/validation of termination of appointment date is feasible without a trip to Archives II.