Fix issue with St. Kitts and Nevis
The <territory-id> child of the root <country-mission> element should correspond to the document’s filename, which should also correspond an entry in the gsh/data/countries-old collection. In other words, <country-mission> should use our legacy system of country IDs—how the printed book grouped countries. If we need to reference a territory from our newer, still-under-development system of country IDs, we place that in the <contemporary-territory-id> column. The entry will still be shown under the legacy country grouping, but the country name in parentheses after the person’s title will use the label from the newer system of country IDs.